+++ src/main/java/net/geforcemods/securitycraft/ClientHandler.java
@@ -5,6 +5,7 @@ import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 import net.geforcemods.securitycraft.blockentities.IMSBlockEntity;
 import net.geforcemods.securitycraft.blockentities.SecretSignBlockEntity;
@@ -76,6 +77,7 @@ import net.minecraft.client.renderer.ItemBlockRenderTypes;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.Sheets;
 import net.minecraft.client.renderer.entity.NoopRenderer;
+import net.minecraft.client.resources.model.BakedModel;
 import net.minecraft.client.resources.model.ModelResourceLocation;
 import net.minecraft.core.BlockPos;
 import net.minecraft.nbt.CompoundTag;
@@ -93,6 +95,7 @@ import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.state.StateHolder;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.client.event.EntityRenderersEvent;
 import net.minecraftforge.client.event.ModelBakeEvent;
@@ -120,22 +123,7 @@ public class ClientHandler {
 
 	@SubscribeEvent
 	public static void onModelBake(ModelBakeEvent event) {
-		String[] facings = {
-				"east", "north", "south", "west"
-		};
 		//@formatter:off
-		ResourceLocation[] facingPoweredBlocks = {
-				new ResourceLocation(SecurityCraft.MODID, "block_change_detector"),
-				new ResourceLocation(SecurityCraft.MODID, "keycard_reader"),
-				new ResourceLocation(SecurityCraft.MODID, "keypad"),
-				new ResourceLocation(SecurityCraft.MODID, "retinal_scanner")
-		};
-		ResourceLocation[] facingBlocks = {
-				new ResourceLocation(SecurityCraft.MODID, "username_logger")
-		};
-		ResourceLocation[] poweredBlocks = {
-				new ResourceLocation(SecurityCraft.MODID, "laser_block")
-		};
 		String[] mines = {
 				"ancient_debris",
 				"blast_furnace",
@@ -167,54 +155,32 @@ public class ClientHandler {
 				"smoker",
 				"stone"
 		};
-		ResourceLocation[] furnaces = {
-				new ResourceLocation(SecurityCraft.MODID, "keypad_furnace"),
-				new ResourceLocation(SecurityCraft.MODID, "keypad_smoker"),
-				new ResourceLocation(SecurityCraft.MODID, "keypad_blast_furnace")
+		Block[] disguisableBlocks = {
+				SCContent.BLOCK_CHANGE_DETECTOR.get(),
+				SCContent.CAGE_TRAP.get(),
+				SCContent.INVENTORY_SCANNER.get(),
+				SCContent.KEYCARD_READER.get(),
+				SCContent.KEYPAD.get(),
+				SCContent.KEYPAD_BLAST_FURNACE.get(),
+				SCContent.KEYPAD_FURNACE.get(),
+				SCContent.KEYPAD_SMOKER.get(),
+				SCContent.LASER_BLOCK.get(),
+				SCContent.PROJECTOR.get(),
+				SCContent.PROTECTO.get(),
+				SCContent.RETINAL_SCANNER.get(),
+				SCContent.SENTRY_DISGUISE.get(),
+				SCContent.TROPHY_SYSTEM.get(),
+				SCContent.USERNAME_LOGGER.get()
 		};
 		//@formatter:on
-		ResourceLocation invScanRL = new ResourceLocation(SecurityCraft.MODID, "inventory_scanner");
-		ResourceLocation projectorRL = new ResourceLocation(SecurityCraft.MODID, "projector");
 
-		for (String facing : facings) {
-			for (ResourceLocation facingPoweredBlock : facingPoweredBlocks) {
-				registerDisguisedModel(event, facingPoweredBlock, "facing=" + facing + ",powered=true");
-				registerDisguisedModel(event, facingPoweredBlock, "facing=" + facing + ",powered=false");
+		Map<ResourceLocation, BakedModel> modelRegistry = event.getModelRegistry();
+
+		for (Block block : disguisableBlocks) {
+			for (BlockState state : block.getStateDefinition().getPossibleStates()) {
+				registerDisguisedModel(modelRegistry, block.getRegistryName(), state.getValues().entrySet().stream().map(StateHolder.PROPERTY_ENTRY_TO_STRING_FUNCTION).collect(Collectors.joining(",")));
 			}
-
-			for (ResourceLocation facingBlock : facingBlocks) {
-				registerDisguisedModel(event, facingBlock, "facing=" + facing);
-			}
-
-			for (ResourceLocation furnace : furnaces) {
-				registerDisguisedModel(event, furnace, "facing=" + facing + ",lit=false,open=false");
-				registerDisguisedModel(event, furnace, "facing=" + facing + ",lit=false,open=true");
-				registerDisguisedModel(event, furnace, "facing=" + facing + ",lit=true,open=false");
-				registerDisguisedModel(event, furnace, "facing=" + facing + ",lit=true,open=true");
-			}
-
-			registerDisguisedModel(event, invScanRL, "facing=" + facing + ",horizontal=true");
-			registerDisguisedModel(event, invScanRL, "facing=" + facing + ",horizontal=false");
-			registerDisguisedModel(event, projectorRL, "facing=" + facing + ",hanging=true");
-			registerDisguisedModel(event, projectorRL, "facing=" + facing + ",hanging=false");
 		}
-
-		for (ResourceLocation poweredBlock : poweredBlocks) {
-			registerDisguisedModel(event, poweredBlock, "powered=true");
-			registerDisguisedModel(event, poweredBlock, "powered=false");
-		}
-
-		ResourceLocation cageTrapRl = new ResourceLocation(SecurityCraft.MODID, "cage_trap");
-		ResourceLocation protectoRl = new ResourceLocation(SecurityCraft.MODID, "protecto");
-		ResourceLocation sentryDisguiseRl = new ResourceLocation(SecurityCraft.MODID, "sentry_disguise");
-
-		registerDisguisedModel(event, cageTrapRl, "deactivated=true");
-		registerDisguisedModel(event, cageTrapRl, "deactivated=false");
-		registerDisguisedModel(event, protectoRl, "enabled=true");
-		registerDisguisedModel(event, protectoRl, "enabled=false");
-		registerDisguisedModel(event, sentryDisguiseRl, "invisible=true");
-		registerDisguisedModel(event, sentryDisguiseRl, "invisible=false");
-		registerDisguisedModel(event, new ResourceLocation(SecurityCraft.MODID, "trophy_system"), "");
 
 		for (String mine : mines) {
 			registerBlockMineModel(event, new ResourceLocation(SecurityCraft.MODID, mine.replace("_ore", "") + "_mine"), new ResourceLocation(mine));
@@ -223,10 +189,10 @@ public class ClientHandler {
 		registerBlockMineModel(event, new ResourceLocation(SecurityCraft.MODID, "quartz_mine"), new ResourceLocation("nether_quartz_ore"));
 	}
 
-	private static void registerDisguisedModel(ModelBakeEvent event, ResourceLocation rl, String stateString) {
+	private static void registerDisguisedModel(Map<ResourceLocation, BakedModel> modelRegistry, ResourceLocation rl, String stateString) {
 		ModelResourceLocation mrl = new ModelResourceLocation(rl, stateString);
 
-		event.getModelRegistry().put(mrl, new DisguisableDynamicBakedModel(event.getModelRegistry().get(mrl)));
+		modelRegistry.put(mrl, new DisguisableDynamicBakedModel(modelRegistry.get(mrl)));
 	}
 
 	private static void registerBlockMineModel(ModelBakeEvent event, ResourceLocation mineRl, ResourceLocation realBlockRl) {