+++ src/main/java/net/geforcemods/securitycraft/blockentities/DisguisableBlockEntity.java
@@ -14,6 +14,8 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.state.properties.BlockStateProperties;
+import net.minecraft.world.level.material.Fluids;
 import net.minecraftforge.client.model.data.IModelData;
 import net.minecraftforge.client.model.data.ModelDataMap;
 import net.minecraftforge.fmllegacy.network.PacketDistributor;
@@ -32,8 +34,14 @@ public class DisguisableBlockEntity extends CustomizableBlockEntity {
 		super.onModuleInserted(stack, module, toggled);
 
 		if (module == ModuleType.DISGUISE) {
-			if (!level.isClientSide)
+			if (!level.isClientSide) {
+				BlockState state = getBlockState();
+
 				SecurityCraft.channel.send(PacketDistributor.ALL.noArg(), new RefreshDisguisableModel(worldPosition, true, stack, toggled));
+
+				if (state.hasProperty(BlockStateProperties.WATERLOGGED) && state.getValue(BlockStateProperties.WATERLOGGED))
+					level.updateNeighborsAt(worldPosition, state.getBlock());
+			}
 			else
 				ClientHandler.putDisguisedBeRenderer(this, stack);
 		}
@@ -44,8 +52,14 @@ public class DisguisableBlockEntity extends CustomizableBlockEntity {
 		super.onModuleRemoved(stack, module, toggled);
 
 		if (module == ModuleType.DISGUISE) {
-			if (!level.isClientSide)
+			if (!level.isClientSide) {
+				BlockState state = getBlockState();
+
 				SecurityCraft.channel.send(PacketDistributor.ALL.noArg(), new RefreshDisguisableModel(worldPosition, false, stack, toggled));
+
+				if (state.hasProperty(BlockStateProperties.WATERLOGGED) && state.getValue(BlockStateProperties.WATERLOGGED))
+					level.getLiquidTicks().scheduleTick(worldPosition, Fluids.WATER, Fluids.WATER.getTickDelay(level));
+			}
 			else
 				ClientHandler.DISGUISED_BLOCK_RENDER_DELEGATE.removeDelegateOf(this);
 		}