+++ src/main/java/net/geforcemods/securitycraft/blocks/ProjectorBlock.java
@@ -52,7 +52,7 @@ public class ProjectorBlock extends DisguisableBlock {
 
 	public ProjectorBlock(Block.Properties properties) {
 		super(properties);
-		registerDefaultState(stateDefinition.any().setValue(FACING, Direction.NORTH).setValue(HANGING, false));
+		registerDefaultState(stateDefinition.any().setValue(FACING, Direction.NORTH).setValue(HANGING, false).setValue(WATERLOGGED, false));
 	}
 
 	@Override
@@ -124,16 +124,12 @@ public class ProjectorBlock extends DisguisableBlock {
 
 	@Override
 	public BlockState getStateForPlacement(BlockPlaceContext ctx) {
-		return getStateForPlacement(ctx.getLevel(), ctx.getClickedPos(), ctx.getClickedFace(), ctx.getClickLocation().x, ctx.getClickLocation().y, ctx.getClickLocation().z, ctx.getPlayer());
-	}
-
-	public BlockState getStateForPlacement(Level level, BlockPos pos, Direction clickedFace, double hitX, double hitY, double hitZ, Player placer) {
-		return defaultBlockState().setValue(FACING, placer.getDirection().getOpposite()).setValue(HANGING, clickedFace == Direction.DOWN);
+		return super.getStateForPlacement(ctx).setValue(FACING, ctx.getPlayer().getDirection().getOpposite()).setValue(HANGING, ctx.getClickedFace() == Direction.DOWN);
 	}
 
 	@Override
 	protected void createBlockStateDefinition(Builder<Block, BlockState> builder) {
-		builder.add(FACING, HANGING);
+		builder.add(FACING, HANGING, WATERLOGGED);
 	}
 
 	@Override